clip/pyproject.toml
[project]
name = "from-scratch"
version = "0.1.0"
description = "clip"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "marimo>=0.20.4",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "tiktoken>=0.12.0",
    "timm>=1.0.25",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "wandb>=0.25.1",
]
